// src/types/vehicle.ts
export type Vehicle = {
  Marca: string;
  Modelo: string;
  Año: number;
  Kilometraje: number;
  Color: string;
  Descripción: string;
  Puertas: number;
  Segmento: string;
  Precio: number;
  Estado: string;
  Ciudad: string;
  "Tipo de combustible": string;
  Motor: number;
  Transmisión: string;
  URL: string;
  Cantidad: number;
};